state.js splitted in more functional separate files, needle working
