+++ src/store/index.js
@@ -1,9 +1,12 @@
 import { createStore } from "vuex"
 import state from "./state"
-
+import daoState from "./dao"
+import marketState from "./market"
 const store = createStore({
   modules: {
-    state
+    state,
+    daoState,
+    marketState
   },
 })
